refactor: Update PrismaModule to conditionally enable logging
The PrismaModule in the prisma module has been updated to conditionally enable logging based on the value of the 'logPrisma' configuration property. If 'logPrisma' is set to 'true', the Prisma logger will be used with the specified environment. Otherwise, logging will be disabled. Additionally, the error format has been set to 'pretty' for improved readability.